requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-community==0.0.10
pydantic==2.5.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==1.4.49
databases==0.8.0
asyncpg==0.29.0
redis==5.0.1
httpx==0.25.2
pillow==10.1.0
google-generativeai==0.3.2
bleach==6.1.0
cryptography==41.0.7